feat(section) : add func registering station on line (adding section)
Input a line, a station and a position, then get a line instance and a station instance if they are in the lists.

And find a section which has the same line instance as inputted and add a station instance to the list of stations in it.

Also needs to add additional functions...
- Check if position is number format
- Add a station to inputted position